requirements.txt
Django
psycopg2-binary
pillow
djangorestframework
python-decouple
djangorestframework-simplejwt
drf-yasg
pillow
django-rest-passwordreset
celery
django-widget-tweaks